Comment 및 Post 인터페이스 구현
comment,post controller 구현하였습니다. 시험삼아 구동하는것때문에 User 쪽 임의로 건드린 부분있습니다(UserId 를 integer 말고 long 으로 변경)

+++ src/main/java/com/delta/delta/controller/UserController.java
@@ -13,7 +13,7 @@ public class UserController {
     UserService userService;
 
     @GetMapping("/{userId}")
-    public User getUser(@PathVariable Integer userId) {
+    public User getUser(@PathVariable Long userId) {
         return userService.getUser(userId);
     }
 
@@ -23,13 +23,13 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public User updateUser(@PathVariable Integer userId, @RequestBody User user) {
+    public User updateUser(@PathVariable Long userId, @RequestBody User user) {
         user.setUserId(userId); // Ensure the userId is set
         return userService.updateUser(user);
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUser(@PathVariable Integer userId) {
+    public void deleteUser(@PathVariable Long userId) {
         userService.deleteUser(userId);
     }
 }
